src/data-monitor/requirements.txt
pymongo==3.7.2
python-dateutil==2.6.0
prometheus_client==0.4.2

